requirements.txt
aiosqlite==0.18.0
alembic==1.10.4
anyio==3.6.2
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
cachetools==5.3.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
databases==0.7.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.92.0
feedparser==6.0.10
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.26.1
packaging==23.1
pandas==2.1.2
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.5
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
sqlparse==0.4.4
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.2.2
uvicorn==0.20.0
